+++ internal/table/session.go
@@ -1374,40 +1374,17 @@ func (s *Session) ReadRows(
 	opts ...options.ReadRowsOption,
 ) (_ result.Result, err error) {
 	var (
-		a       = allocator.New()
-		request = Ydb_Table.ReadRowsRequest{
-			SessionId: s.id,
-			Path:      path,
-			Keys:      value.ToYDB(keys, a),
-		}
+		a        = allocator.New()
+		request  = makeReadRowsRequest(a, s.id, path, keys, opts)
 		response *Ydb_Table.ReadRowsResponse
 	)
 	defer func() {
 		a.Free()
 	}()
 
-	for _, opt := range opts {
-		if opt != nil {
-			opt.ApplyReadRowsOption((*options.ReadRowsDesc)(&request), a)
-		}
-	}
+	response, err = s.client.ReadRows(ctx, request)
 
-	response, err = s.client.ReadRows(ctx, &request)
-	if err != nil {
-		return nil, xerrors.WithStackTrace(err)
-	}
-
-	if response.GetStatus() != Ydb.StatusIds_SUCCESS {
-		return nil, xerrors.WithStackTrace(
-			xerrors.FromOperation(response),
-		)
-	}
-
-	return scanner.NewUnary(
-		[]*Ydb.ResultSet{response.GetResultSet()},
-		nil,
-		scanner.WithIgnoreTruncated(s.config.IgnoreTruncated()),
-	), nil
+	return makeReadRowsResponse(response, err, s.config.IgnoreTruncated())
 }
 
 // StreamExecuteScanQuery scan-reads table at given path with given options.