+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added `table.Client.ReadRows` method with internal retries
+
 ## v3.100.3
 * Fixed bug with concurrent rewrites source slice of `grpc.DialOption` on dial step